перевод на gcc

+++ library/config.h
@@ -1,10 +1,12 @@
 #ifndef CONFIG_H
 #define CONFIG_H
 
+#include <stdio.h>
+
 static const size_t STD_STACK_CAP = 10;
 static const size_t MAX_SIZE_STR  = 100;
 static const char   VENOM_NUM_CMD = -13;
-static const int    LEN_NAME_REG  = 3;
+static const size_t LEN_NAME_REG  = 3;
 
 typedef double num_t;
